requirements.txt
pigpio>=1.78
smbus2>=0.4.2
mpu9250-jmdev>=1.3
bmp280>=0.1.0
gpsd-py3>=0.3.0
pyserial>=3.5
opencv-python-headless>=4.8.0
numpy>=1.22 